refactor: prefer type suffix instead of as conversion, remove extra asserts

+++ crates/interledger-btp/src/packet.rs
@@ -96,9 +96,9 @@ where
     let mut protocol_data = Vec::new();
 
     let num_entries = reader.read_var_uint()?;
-    let mut i = BigUint::from(0 as u32);
+    let mut i = BigUint::from(0u32);
     while i < num_entries {
-        i = i.add(BigUint::from(1 as u8)); // this is probably slow
+        i = i.add(BigUint::from(1u8)); // this is probably slow
         let protocol_name = String::from_utf8(reader.read_var_octet_string()?)?;
         let content_type = ContentType::from(reader.read_u8()?);
         let data = reader.read_var_octet_string()?;
